WIT working well now with widget
